remove route

+++ app/controllers/activities/mountain_routes_controller.rb
@@ -23,7 +23,7 @@ module Activities
 
     def destroy
       route = Db::Activities::MountainRoute.find(params[:id])
-      fail 'not admin' unless user_signed_in? && current_user.admin?
+      fail 'not admin' unless user_signed_in? && (current_user.admin? || current_user.id == route.user_id)
       route.destroy
 
       redirect_to activities_mountain_routes_path, notice: t('.removed_successfully')
